# perf: Use (slightly) optimized cache key calculation algorithm