--- a/src/decoder.rs
+++ b/src/decoder.rs
@@ -90,8 +90,18 @@ impl Decoder {
         if pc as usize >= RISCV_MAX_MEMORY {
             return Err(Error::MemOutOfBound);
         }
-        // according to RISC-V instruction encoding, the lowest bit in PC will always be zero
-        let instruction_cache_key = (pc >> 1) as usize % INSTRUCTION_CACHE_SIZE;
+        let instruction_cache_key = {
+            // according to RISC-V instruction encoding, the lowest bit in PC will always be zero
+            let pc = pc >> 1;
+            // Here we try to balance between local code and remote code. At times,
+            // we can find the code jumping to a remote function(e.g., memcpy or
+            // alloc), then resumes execution at a local location. Previous cache
+            // key only optimizes for local operations, while this new cache key
+            // balances the code between a 8192-byte local region, and certain remote
+            // code region. Notice the value 12 and 8 here are chosen by empirical
+            // evidence.
+            ((pc & 0xFF) | (pc >> 12 << 8)) as usize % INSTRUCTION_CACHE_SIZE
+        };
         let cached_instruction = self.instructions_cache[instruction_cache_key];
         if cached_instruction.0 == pc {
             return Ok(cached_instruction.1);