This new way of cache key calculation could help reduce the number of instructions decoded by ckb-vm. For secp256k1 benchmarks, decoder cache hit ratio increases from 34% to 41%, while the raw number of instructions being decoded drops from 14445 to 12925. For halo2 sample, decoder cache hit ratio increases from 45% to 56%, while raw number of decoded instructions drops from 114462 to 90639 in my testing.

However, the raw performance is not affected as much, which is most likely due to the fact that decoding time contributes to only a small fraction of the overall runtime. The following results can be noticed when benching secp256k1 contracts:

```
interpret secp256k1_bench via assembly
                        time:   [2.6053 ms 2.6065 ms 2.6078 ms]
                        change: [-3.7691% -3.6936% -3.6191%] (p = 0.00 < 0.05)
                        Performance has improved.

interpret secp256k1_bench via assembly mop
                        time:   [2.6798 ms 2.6814 ms 2.6832 ms]
                        change: [-3.0029% -2.7658% -2.5892%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Tho it must be mentioned that secp256k1 benchmarks have a very high variance, chances are the results obtained here are due to CPU frequency changes.

We also noticed a small gain when running halo2 contracts(~133ms to ~129ms). Tho the variance is still high enough that the results here are quite flaky.

One thing we know for sure, is that the new key calculation algorithm will certain help reduce the raw number of instructions being decoded.